Nested app auth implementation (#2168)
- Added parameters for brkRedirectId and brkClientId to pass in /auth
and /token requests for NAA request
- Min required protocol version for NAA is 15
- Made changes to NOT save tokens when the flow is an NAA request. We
have designed NAA flow to not save tokens as OneAuth is anyway saving
the AT acquired and requests broker only when AT is expired. When AT is
expired, we anyway have to make a fetch AT with PRT or RT request. So,
saving the tokens in the cache would not be useful

Design PR link :
https://identitydivision.visualstudio.com/DevEx/_git/AuthLibrariesApiReview/pullrequest/7876