Merged PR 10002: PR feedback: Sign In
Added licenses to missing files, added Javadoc to Sign In classes

Note: I've added Javadoc to classes and methods that are 'new' to our architecture ie: SignInInteractor, Requests, Responses, and Result classes. There are certain classes such as Command Parameters that won't have documentation on them, but that is the case for existing Command Parameters in MSAL currently.

Related work items: #2701158