Base work for the CIAM features (#2166)
The code in this pull request is for the base work needed to support Signup, SignIn and SSPR features for Native Auth. Any class that is used by at least two of the features has been included in this pull request.

This pull request introduces `mockk` and `mockito` libraries as dependencies for test projects. The `common` library test project has a new dependency on `testutils` library project.

A companion PR in the MSAL repository is
https://github.com/AzureAD/microsoft-authentication-library-for-android/pull/1897