AbstractFidoChallengeHandler and telemetry names (#2183)
### Summary
This PR adds the classes of AbstractFidoChallengeHandler,
~~ITelemetryHelper, IFidoTelemetryHelper, and FidoTelemetryHelper~~ and
span and attribute names for future telemetry calls.
The eventual PasskeyFidoChallengeHandler, which will take care of
getting the FIDO assertion and sending it to the server, will extend
AbstractFidoChallengeHandler.
~~ITelemetryHelper is an interface which helps consolidate the setStatus
methods of the CBA and FIDO telemetry helpers. IFidoTelemetryHelper has
methods to set the Challenge and ChallengeHandler names on the future
Fido span. Noting these object names will help determine the type of
FIDO request (authentication vs registration) and the type of passkey
used (regular passkey on device, or security key). The main
implementation of IFidoTelemetryHelper is FidoTelemetryHelper, and a
future implementation will be done for unit testing purposes.~~
None of these classes are currently being used in logic, but reviewing
them now helps cut down the future PR which will contain the
PasskeyFidoChallengeHandler implementation.

### Related PRs
- https://github.com/AzureAD/ad-accounts-for-android/pull/2547