[E2E Test] Fix OneAuthTestApp.handleFirstRun (#2206)
The current implementation logic of OneAuthTestApp.handleFirstRun is:
switch the config first, then turn on "Prefer AAD Broker" flag. However,
OneAuthTestApp decides whether to truly enable the Broker when switching
the config. Thus, the current order doesn't activate the Broker as
expected.

The solution is to turn on "Prefer AAD Broker" before switching the
config.