Passing span context to AuthorizationActivity (#2209)
### Summary
Currently, CBA spans are not correlating correctly with their parent
span (supposed to be AcquireTokenInteractive). A new trace_id is being
created for CBA spans upon instantiation, when we want the trace_id from
the AcquireTokenInteractive span to be populated there instead.
This is because the span context doesn't get inherited when we start the
AuthorizationActivity, the main activity where the CBA code (and much
logic running behind the embedded WebView) resides. The workaround for
this is to put a SerializableSpanContext instance as an extra into the
intent that starts AuthorizationActivity. The context is then retrieved
and saved as a class variable within the Activity. The CBA classes can
reference the span context as a parent span when it creates its own span
(which means they'll share the same trace_id).