Fix lockfile path in internal VSTS pipeline (#2201)
Internal VSTS pipeline is broken after the android gradle plugin version
was updated in the last release. We did not catch this in last release
because we used External Maven pipeline. But, this time, since we are
doing a hotfix, we used the internal VSTS pipeline and it failed because
lockile path was not updated.
https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1184694&view=results

Successful pipeline run with my fix :
https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1185083&view=logs&j=10e7b5b4-5c92-50ca-fdf6-60562150c652
Note : This run is failing at a different stage (Publish to VSTS stage,
but the generate dependencies lockile and Publish Artifact: dependencies
lockfile tasks are passing.

Here is an old PR I raised for fixing similar issue in daily pipeline
https://github.com/AzureAD/android-complete/pull/248/files