Merged PR 10401: Use CharArray/char[] for password
Use CharArray/char[] for password

This PR introduces a change for password signature to mitigate an issue with JVM strings.

JVM strings are immutable, therefore any data stored in any String will stay in the memory until it is garbage collected. In addition to that, after the memory is deallocated and released by the application using MSAL, it is possible for another application is given the ownership of the same memory area after which point the app has a chance of recovering the value of the String by doing a memory sweep. This is a highly advanced (figuring out which part of the memory is password is not straightforward) and mostly up to chance (the app might not be given the same memory space) security attack method, but it is possible and is a previously observed attack method. With using a `CharArray` in Kotlin or `char[]` in Java which are mutable, the memory can be reset to another value before being garbage collected or deallocated which makes this attack not practically possible (there are other ways to recover memory from physical disk, but it's out of scope of this action item).

This PR does not aim to mitigate the issue completely, as the password being stored will be serialized into string before being sent in a request. To completely mitigate the issue, the infrastructure and dependencies used to send the requests need to be changed, but it is a bigger effort involving most part of existing MSAL, which cannot be done in a single PR. Therefore, the aim here is to make the necessary changes in the interface so the attack surface is reduced, and the interface won't need to be updated when it is decided to completely mitigate the issue (a minor version update will be enough which would be a breaking change with a major update otherwise).

This PR includes changes in the common.

Related work items: #2722916