Catch unknown Keystore errors during decryption (#2255)
Issue :
https://portal.microsofticm.com/imp/v3/incidents/incident/443725136/summary
In the above IcM, we are seeing some unknown keystore errors during
keypair generation. This is only happening when KS is getting accessed
during app boot up. we do not have a repro for the issue and it is
happening on 0.01% of devices. We have considered 3 ways of fixing this.

1. Find the actual root cause of the issue. For this, we have created a
bug on google. This may take some time and we would need to come up with
a short term solution
2. Since the KS errors are only happening on Android 10 and 11, we can
setBrokerDiscovery as false on these versions in OneAuth. Howver, this
would need to be reverted when we need to enable broker discovery for
LTW.
3. Catch the KS errors and throw a client exception like we did in this
PR
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2190.