Handle Receiver Callback Exception in CommandDispatcher interactive command (#2305)
Related Bug
https://identitydivision.visualstudio.com/Engineering/_workitems/edit/2502607

Related PR https://github.com/AzureAD/ad-accounts-for-android/pull/2690

Teams is seeing a crash during
`LocalMSALController.onFinishAuthorizationSession()`, where we
occasionally get a `NullPointerException` on
`mAuthorizationStrategy.completeAuthorization()`. It's unclear why this
happens, but the root issue that such an exception should be routed as a
BaseException and sent back as a command result. This PR handles this by
routing the exception through the command result if it happens during
the callback

Also add a check to see if the authorization future is initialized and
waiting, provide an exception for it if we run into the NPE.

E2E testing
I tested this with MSALTestApp by throwing a NPE explicitly in
complereAuthorization() and caught the exception, routing through to a
command result.