Revert changes made to method signature for futures (#2307)
Revert changes made to method signature to return back a ResultFuture
instead of a Future. It seems to have introduced a breaking change in
MSAL, instead of debugging this during release time, I rolled back part
of the solution for the original google crash. The fix still does the
same thing, just by using a type casting rather than explicit return
type being set to ResultFuture. As mentioned in the original PR, this
scenario should theoretically be impossible anyway, as if we fail out of
authorization because the strategy is null, then the future would not be
initialized, but keeping the result future check there for additional
security.

Original PR
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2305