Revert changes made to method signature for futures (#2309)
Revert changes made to method signature to return back a ResultFuture
instead of a Future. It seems to have introduced a breaking change in
MSAL, instead of debugging this during release time, I rolled back part
of the solution for the original google crash. In this PR i'm rolling
back the whole fix for better tracking, will have a following PR to
reintroduce the smaller fix into common to address the issue without
adjusting method signatures that could affect msal or broker.

Original PR
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2305