Gracefully handles unexpected error in encryption layer (#2318)
## Why?

We're getting unexpected error from KeyStore error from 14k Office app
users. (~0.02%).

https://identitydivision.visualstudio.com/Engineering/_workitems/edit/2835735

The stack trace implies that this is a OS-level issue. Somehow keystore
is not working properly if it's invoked when the app is being booted up.

We've filed a fix to Google, but it hasn't been picked up yet.

On our side, we should be handle these errors gracefully, not crashing
the hosting apps.

## Changes
1. Capture all unexpected throwable as ClientException with
UNKNOWN_CRYPTO_ERROR.
ClientException is not a RuntimeException and will be handled by the
caller.

2. Catch the unexpected exception in AndroidKeyStoreUtil, as seen in the
stack trace.

3. In decrypt(), only wraps error when there is more than one exception.
(multiple key is used to decrypt, which is not the case for
ADAL/MSAL/OneAuth)

I also removed isEncryptedByThisKeyIdentifier() because it's redundant
with getKeyIdentifierFromCipherText().
This method is being consumed by broker, therefore it's a breaking
change.
Related Broker PR:
https://github.com/AzureAD/ad-accounts-for-android/pull/2710