Passkey error prefix (#2336)
### Summary
A prefix to errors sent via the passkey protocol is being added. This
prefix allows ESTS to recognize that we're not sending them a successful
assertion. The handling of these errors for now is for ESTS to redirect
back to the main login page, but this design will be changed later to
include more helpful error messages.