getIpcStrategies should only be invoked in bg thread (#2339)
OneAuthSharedFunctions.getIpcStrategies, which invokes
ContentProviderStrategy.isSupportedByTargetedBroker, which invokes IPC
operation, should be invoked in background thread only.

The issue is that we're invoking it in the constructor today - and that
constructor is invoked in main thread.

by making this lazy-load, it would only be invoked when the ipc request
is actually happening - inside a command, which sits inside a
commandDispatcher.