Pipelines fixes. (Credscan suppression file, CodeQL, AndroidAuthClientVariables) (#2382)
1. Credscan suppression file
In order to be compliant with EO we move the production pipelines to 1ES
Pipeline Templates, this templates auto-inject some sdl tasks like
credscan, that scan all the repos used.
In this case credscan found a couple of 'vulnerabilities' that blocks
the pipeline, in order to ignore these false alarms, we need to include
this file.
2.  CodeQL
Semmle guardian task unexpected stop working. Probably because this task
is on deprecation path.
https://eng.ms/docs/cloud-ai-platform/devdiv/one-engineering-system-1es/1es-docs/codeql/troubleshooting/onboarding/guardian
For this reason, we move to CodeQL3000
3. Stop using AndroidAuthClientVariables and use
AndroidAuthClientAutomationSecrets instead
The service connection for this group variable was deactivated, instead
of activating it, I decided to remove it since it has low usage, and
decided to move the only valid secret to another group variable.