Implement new Temp User Creation and reset APIs (#2421)
Recently the TempUserCreation functionality of labs api was flagged for
deletion due to security concerns. I've collaborated with Gladwin, along
with OneAuth and iOS teams, to help implement a new API to replace our
usage of the original api for temp user creation. This PR includes
changes to implement the new API into our testing libraries.