Platform Specific Extra Query Parameters (#2426)
### Summary
Adding a method to IPlatformUtil
`setPlatformSpecificExtraQueryParameters`. This method takes in an
existing list of parameters or null and returns a list of parameters
which has been edited based on the platform, or null (if the existing
list was null).
An example of a query string parameter which benefits from this method
is the passkey parameter, which only needs to be sent for Android broker
and not Linux broker.

This is going to make the consumer checks fail at the broker stage, but
I'll run them all once to make sure that the other libraries don't run
into issues.