Support Passing x-client-xtra-sku header to ESTS only for brokered requests (#2430)
Feature Spec:
https://microsoft-my.sharepoint-df.com/:w:/g/personal/oldalton_microsoft_com1/ET9igW1SaHJLmlNIBi_CbqoBuyAKPWIZ0BpzcxTUB8yeLw?e=xmY1xe&nav=eyJoIjoiMTk3NDU4ODU3MyJ9&ovuser=72f988bf-86f1-41af-91ab-2d7cd011db47%2Cfadidurah%40microsoft.com&clickparams=eyJBcHBOYW1lIjoiVGVhbXMtRGVza3RvcCIsIkFwcFZlcnNpb24iOiI0OS8yNDA1MzEwMTQwNyIsIkhhc0ZlZGVyYXRlZFVzZXIiOmZhbHNlfQ%3D%3D

Changes to enable passing in the new x-client-xtra-sku header to ests
when calling /authorize and /token endpoints. This field has a standard
schema agreed upon by all teams JavaDoc of `ClientExtraSkuAdapter.java`
explains the schema of this new field, most of the fields in this new
file.

To facilitate this, I'll need to propogate Product and Porduct version
through the authorization workflow, as it runs on a new thread, and the
DiagnosticContext wasn't being populated. I've updated the bundle to
pass in these values so we can use them in the telemetry field, also
used them to populate the diagnostic context for good measure.

For /Token endpoint, I noticed that in some cases the broker version
wasn't being passed, even if it was a brokered scenario. This didn't
seem to be by design, so i updated it to add the broker version in these
flows. Please reference Broker PR
(https://github.com/AzureAD/ad-accounts-for-android/pull/2832). This is
important as i'm using the presence of the broker version as a check for
sending in the new telemetry field, as we only want to send it in
brokered scenarios.


Validated this in interactive and silent requests for the following
apps:
OneAuth (MSA account)
MsalTestApp
AdalTestApp

---------

Co-authored-by: Shahzaib <37125644+shahzaibj@users.noreply.github.com>