Add check for unset correlation ID when configuring request headers (#2435)
If a correlation ID hasn't been configured for a request, the 'UNSET'
value will be sent to the server, and the server will then return that
'UNSET' value from the response. To prevent this, the correlation ID
should only be appended if it is a valid value.

Additionally, tests have been added to ensure that an 'UNSET' value
isn't added to the request headers, and valid correlation IDs are being
added to the request headers on a valid request.

Also, there are a few tweaks made to `NativeAuthRequestHandlerTest` -
there are some tests where a request is configured and created, but
there are no assertions made on those requests. I've added a few checks
to tests that were missing validations, and some other small tweaks.

Update: 
“If a correlation ID hasn't been configured for a request, the 'UNSET'
value will be sent to the server, and the server will then return that
'UNSET' value from the response.” -> After experiment, this assumption
is wrong. The header used on the SDK side is client-request-id not
ms-client-request-id, these two values are different if they both
attached in the response header. The interesting thing is that, if
client sends invalid correlation id like "UNSET", the server would not
return the client-request-id to the client. On the contrary, if the
correlation id is valid and client sent it in the client-request-id
header, the server will return it as the same value in the
client-request-id header. SDK made the assumption that if no
client-request-id header is sent to the server from client, the server
will generate one for the client seems untrue. The correlation id is
dependant on client-request-id x-ms-request-id is the request specific,
it targets the Service Request Id not the CorrelationId, while
x-ms-request-id is returned from the server in all times.

---------

Co-authored-by: yuxin <yuxin@microsoft.com>