Moving OS version check for passkeys (#2450)
### Summary
If a device's OS does not support passkeys, then we shouldn't even
reveal the passkey option. This means that the OS check should be done
while we are deciding to add `webauthn=1` or not.
As a result of needing to move the OS checks to earlier points
(specifically when we add the extra query param `webauthn=1`), we also
saw that the placement of the passkey extra query param logic needed to
be revised. The `updateWithAndGetPlatformSpecificExtraQueryParameters`
implementation for `AndroidPlatformUtil` is cleared and is instead put
in the implementation of `AndroidBrokerPlatformUtil`, a new class which
can be found in this PR:
https://github.com/AzureAD/ad-accounts-for-android/pull/2858/files

### Related PRs
- https://github.com/AzureAD/ad-accounts-for-android/pull/2858
-
https://github.com/AzureAD/microsoft-authentication-library-for-android/pull/2138