MFA error handling (#2493)
When MFA is required during a call to /token with
grant_type=refresh_token, we need some native auth specific handling to
inform developers on how to handle this error.

**Note**: Ideally I'd like to store the `50076` error code in a
dedicated file, but I'm unsure what to call such a file, as
`OAuth2ErrorCode` already exists. IMO this class name is wrong, as these
are OAuth2 errors, not error codes.