Use legacy FIDO2 API to provide security key support to Android 13 devices, Fixes AB#3027082 (#2508)
### Summary
Android's [Credential Manager
API](https://developer.android.com/identity/sign-in/credential-manager)
does not have FIDO2 security key support for Android versions 13 and
below. They recommend using their [legacy FIDO2
API](https://developers.google.com/identity/fido/android/native-apps)
instead. Thus, this PR includes and uses the legacy FIDO2 API in order
to provide FIDO2 security key support for our customers using Android 13
and below.

As discussed with the PMs, we are going to use this API as a fallback
from an attempt with CredMan in the case where a `NoCredentialException`
is thrown AND the device OS is Android 13 or lower.
As a supplement, we're setting `preferImmediatelyAvailableCredentials`
in the CredMan `getCredentialRequest` object to true if the device OS is
Android 13 or lower. This ensures the behavior where no dialog from
CredMan is shown if no passkey cred is present. The end goal is for an
Android <= 13 user who only has a security key to see one dialog which
will allow them to authenticate.

This PR includes some new classes in the `fido` package. I attempted to
draw a diagram to explain the relation:

![image](https://github.com/user-attachments/assets/78e08d37-e5ac-4cc3-abb6-593e666bdbed)

The legacy FIDO2 API provides a `PendingIntent` which the host app needs
to start, and when the resulting Activity closes is when we get our
results. This is different from CredMan's exposed endpoints, and thus I
tried to wrap the legacy API in such a way which mimics the behavior of
CredMan.

Some unit tests have been added, and I am going to continue to add unit
tests this month, but planning on splitting this up into another PR
since this PR is already on the longer side. (Shadow and extended
classes will be needed)

There is a flight variable guarding the feature, but since the logic is
in common, the `CommonFlight` variable is set to true by default, while
the `BrokerFlight` variable is set to false by default.



[AB#3027082](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3027082)