Translate MFA token error to UIRequiredException instead of ServiceException to release branch, Fixes AB#3090086 (#2540)
When MFA is required, the token error response was previously translated
to a ServiceException instead of a UIRequiredException. With this
update, the token error will be translated to a UIRequiredException when
MFA is required.
Customization of native authentication error descriptions will only
occur when the native authentication interface is being used.

[AB#3090086](https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3090086)