Removing unnecessary attributes from keystore wrap operation, Fixes AB#3137956 (#2578)
In an older
[PR](https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2558),
I replaced the deprecated Keystore API for Android 28+
But it did not fix the issue of unwrap keystore operation failing on
some Pixel 5 devices. Since it is not reproducing on other devices, I am
removing the unnecessary attributes from the keystore wrap operation to
see if it fixes the issue. certificate parameters are no longer actively
used in other parts of the code. Hence removed those.

I have already shared the test build with 2 customers and the issue has
NOT reproduced on their devices so far with this fix.

Fixes
[AB#3137956](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3137956)