Fix multiple prompts issue in cross cloud request, Fixes AB#3179147 (#2599)
**Issue** :
https://portal.microsofticm.com/imp/v5/incidents/details/603998142/summary
When cross cloud authentication happens, webview could be redirected to
a different cloud rather than the initial cloud loaded in the webview.
When such a redirect happens, the request have no more PRT header
attached that can be consumed by the new cloud. Therefore users will be
asked to input credential when redirected to a new cloud.

**Fix** : In this PR, we are adding the support to attach cross domain
PRT in embedded webview requests.
To improve the SSO experince, we are now making the change to detect
redirects between clouds and attach available PRT for the new cloud if
cross cloud authentication happens.

**Testing** : 
For testing the fix
1. Ran all UI tests to ensure existing tests are not breaking
https://identitydivision.visualstudio.com/Engineering/_build/results?buildId=1442539&view=ms.vss-test-web.build-test-results-tab&runId=4591284&resultId=100022&paneView=debug
2. Ran below steps to repro the issue and fix it

- Install MsalTestApp and BrokrHost
- In MsalTestApp, use default config, loginhint
[idlabxcg@msidlab4.onmicrosoft.com](mailto:idlabxcg@msidlab4.onmicrosoft.com)
and deviceId claim attached. Complete the sign in.
- Next, change the config to Arlington, loginhint
[idlabxcg@msidlab4.onmicrosoft.com](mailto:idlabxcg@msidlab4.onmicrosoft.com),
deviceId claim attached, authority
https://login.microsoftonline.us/45ff0c17-f8b5-489b-b7fd-2fedebbec0c4
and click on acquire token
- Without my fix, we would be prompted to enter password twice at this
point.
- With my fix, we are no longer prompted to enter password at all.

Fixes
[AB#3179147](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3179147)