Corrected error handling in cross cloud scenario, Fixes AB#3179147 (#2602)
In this PR, we fixed multiple prompt issue in cross cloud scenarios
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2599
But when an exception is thrown while loading PRT or when getting nonce,
we are not handling the exception correctly. The expected case should be
that the exception is ignored and the webview should be loaded
correctly.
What is actually happening is, the webview stops loading.

Fixes
[AB#3179147](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3179147)