Only show QR +PIN rationale when the request origin is from Microsoft (#2613)
This change here is a temporary fix for QR + PIN.
The final solution will be incorporated in the next month April 2025.
On the final solution we only allow QR + PIN prompt when the
Authenticator RestrictionsManager contains QR PIN key.

This fix will limit the QR +PIN camera consent prompt to requests coming
from login.microsoftonline.com only.
We also reworked the CameraPermissionRequest to avoid NPR when
PermissionRequest is null, by calling only the
CameraPermissionRequestHandler that is null safe.

---------

Co-authored-by: melissaahn <melissaahn@microsoft.com>
Co-authored-by: Azure DevOps Pipeline <build-agent@azuredevops.com>
Co-authored-by: Sowmya Malayanur <69237821+somalaya@users.noreply.github.com>