Native auth: network layer for JIT feature, Fixes AB#3188135 (#2626)
Adding network layer for JIT feature. Integration tests are written for
the NativeAuthOAuth2Strategy class. To write tests for this class, I
need to create new commandParameters classes. Since these classes are
more related to the business logic, I decided to add them and the
integration tests in the next PR.


[AB#3188135](https://identitydivision.visualstudio.com/Engineering/_workitems/edit/3188135)