Seems UI is still hopping between legacy msa ui and fluent ui, changes to support both, Fixes AB#3246594 (#2634)
It seems that the fluent ui at the time of writing this PR (4/29/2025)
is still not fully complete, some manual testing showed that the legacy
ui and fluent ui are still appearing at random points. This PR will make
a change using regex to support both UI's and pass the test regardless
of which UI we get.


[AB#3246594](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3246594)