Expose a JavaScript API in brokered Webviews to facilitate Improved Same Device NumberMatch , Fixes AB#3203956 (#2617)
SPEC:
https://microsoft-my.sharepoint-df.com/:w:/r/personal/siddhijain_microsoft_com/_layouts/15/Doc.aspx?sourcedoc=%7BD1D944D5-2047-40AB-B8F9-98506BF212A3%7D&file=Engineering%20design%20-%20Number%20matching%20on%20same%20device.docx&action=default&mobileredirect=true&share=IQHVRNnRRyCrQLj5mFBr8hKjAbj81fEnpO6X-99laqs2j_M&wdOrigin=TEAMS-MAGLEV.p2p_ns.rwc&wdExp=TEAMS-TREATMENT&wdhostclicktime=1743094076241&web=1

Word Doc: for JavaScript Api:
https://microsoft-my.sharepoint-df.com/:w:/p/veenasoman/EY1AZIeT8X5KrXVz97Vx520B3Jj0fBLSPlklnoRvcmbh0Q?e=ZVVUrw&nav=eyJoIjoiMjEzMzE1Mzg5NSJ9

Structure has changed a bit for this. To facilitate future work, we will
have a generalized JavaScript API that takes in a json string payload.
This is used to parse out a function name, and data field, both of which
are used to call a specific function in broker code. This same
functionality will be used next month for CA Block improvment work (I
don't have a spec to this one yet).

Expected method call in JavaScript is now something like this, we are
working on finalizing json schema:
`BrokerJS.postToBroker('{function: NUMBER_MATCH,data: {sessionID: id,
numberMatch: number}}')`

I added some unit tests in the broker PR, but primary validation will be
when ests exposes a test slice that calls the JavaScript API. Did some
testing in our webview class to call javascript code, and was able to
prompt the numberMatch method.

Broker PR: https://github.com/AzureAD/ad-accounts-for-android/pull/3073


[AB#3203956](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_workitems/edit/3203956)